Move DATABASE_URL resolver test
